+++ next.config.mjs
@@ -1,13 +1,7 @@
 import createMDX from "@next/mdx";
-import remarkMath from "remark-math";
-import rehypeKatex from "rehype-katex";
 
 const withMDX = createMDX({
   extension: /\.mdx$/,
-  options: {
-    remarkPlugins: [remarkMath],
-    rehypePlugins: [rehypeKatex],
-  },
 });
 
 /** @type {import('next').NextConfig} */